requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.0
defusedxml==0.8.0rc2
dj-database-url==2.2.0
Django==5.0.7
django-cors-headers==4.4.0
django-filter==24.3
django-jazzmin==3.0.0
django-rest-knox==5.0.1
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
gunicorn==19.9.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Markdown==3.6
MarkupSafe==2.1.5
mysqlclient==2.2.4
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.1
pillow==10.4.0
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
setuptools==72.2.0
simplejson==3.19.2
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
whitenoise==6.7.0
